+++ phidgets_digital_inputs/include/phidgets_digital_inputs/digital_inputs_ros_i.hpp
@@ -58,7 +58,7 @@ class DigitalInputsRosI final : public rclcpp::Node
 
     void timerCallback();
     rclcpp::TimerBase::SharedPtr timer_;
-    int publish_rate_;
+    double publish_rate_;
 
     void publishLatest(int index);
